+++ ACID/Assets/Prefabs/Effects/HurtFlash.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b719f7402bcb7414d9042a9633007180
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
