+++ ACID/Assets/Prefabs/Effects/Image.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19b164221664f42cea3e40849ca8f0eb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
