+++ ACID/Assets/Prefabs/Effects/JustFine.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0df475c5a72b4f70a6a739011bfa67d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
